fix(solutions): memory issues thanks to -fsanitize=address flag with gcc

+++ challenges/fibonacci/solutions/c/function/solution.c
@@ -1,13 +1,13 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int fibonacci(int number) {
+unsigned long fibonacci(unsigned long number) {
   return number < 2 ? number : fibonacci(number - 1) + fibonacci(number - 2);
 }
 
 int main() {
-  int number;
-  scanf("%d", &number);
-  printf("%d\n", fibonacci(number));
+  unsigned long number;
+  scanf("%lu", &number);
+  printf("%lu\n", fibonacci(number));
   return EXIT_SUCCESS;
 }
