+++ challenges/is-prime-number/solutions/c/function/solution.c
@@ -2,8 +2,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-bool is_prime_number(int number) {
-  for (int iteration = 2; iteration < number; iteration++) {
+bool is_prime_number(unsigned long number) {
+  for (unsigned long iteration = 2; iteration < number; iteration++) {
     if (number % iteration == 0) {
       return false;
     }
@@ -12,8 +12,8 @@ bool is_prime_number(int number) {
 }
 
 int main() {
-  int number;
-  scanf("%d", &number);
+  unsigned long number;
+  scanf("%lu", &number);
   bool is_prime = is_prime_number(number);
   printf("%s\n", is_prime ? "true" : "false");
   return EXIT_SUCCESS;
